Work on Decoupling Metric and Goal objects.  Introduced Tracker object to compose Goal and Metric.  Layers now work in terms of Tracker instead of Metric.

+++ src/main/java/com/adarga/domain/Goal.java
@@ -18,23 +18,23 @@ public class Goal {
         this.description = description;
     }
 
-    private int goal;
+    private float goal;
     private String title;
     private String description;
 
     public Goal() {}
 
-    public Goal(int goal) {
+    public Goal(float goal) {
         this.goal = goal;
     }
 
-    public Goal(int goal, String title, String description) {
+    public Goal(float goal, String title, String description) {
         this.goal = goal;
         this.title = title;
         this.description = description;
     }
 
-    public int getGoal() {
+    public float getGoal() {
         return goal;
     }
